Move common utils into file, unique bot names
